get skills by category

+++ controllers/skills.js
@@ -6,20 +6,33 @@ const getSkills = async (req, res) => {
   const skills = await DB.find().populate("skillsCat", "name -_id");
   Helper.fMsg(res, "All Skills", skills);
 };
+const getSkillsByCat = async (req, res) => {
+  const skills = await DB.find({ skillsCat: req.params.catId });
+  Helper.fMsg(res, "Skills by CatId", skills);
+};
 const addSkill = async (req, res) => {
   let skill = await new DB(req.body).save();
-  Helper.fMsg(res, "Add new Skill", skill);
+  let addedSkill = await DB.findById(skill._id).populate(
+    "skillsCat",
+    "name -_id"
+  );
+  Helper.fMsg(res, "Add new Skill", addedSkill);
 };
 const deleteSkill = async (req, res) => {
   let skill = await DB.findByIdAndDelete(req.params.id);
-  deleteFile(skill.image);
-  Helper.fMsg(res, "Delete skill", skill);
+  if (skill) {
+    deleteFile(skill.image);
+    Helper.fMsg(res, "Delete skill", skill);
+  }
 };
 const updateSkill = async (req, res, next) => {
   let skill = await DB.findById(req.params.id);
   if (skill) {
     await DB.findByIdAndUpdate(skill._id, req.body);
-    let updatedSkill = await DB.findById(skill._id);
+    let updatedSkill = await DB.findById(skill._id).populate(
+      "skillsCat",
+      "name -_id"
+    );
     Helper.fMsg(res, "update skill", updatedSkill);
   } else {
     next(new Error("Error , No Social with that id"));
@@ -30,4 +43,5 @@ module.exports = {
   addSkill,
   updateSkill,
   deleteSkill,
+  getSkillsByCat,
 };
